+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 ## [Unreleased]
 
 ### Added
+
 - Copy Interpolated String value intention
+- DSL builder generator
+- Update shortcuts for wrap with coroutine builders
 
 ## [2.0.1] - 2026-04-04